Exercicios1_24/Exercise_15/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-queue
pyodbc
requests    
azure-storage-blob
uuid



